// migrations/snapshots/3e59e30f6c4a1f0af569482c391acf861c88c70a0f5cf1940d06c526b0a18938/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly AuditLogs: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly taskId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly changedColumn: Varchar<100>;
      readonly oldValue: CodecTypes['pg/text@1']['output'] | null;
      readonly newValue: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly ProjectMembers: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly projectId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Projects: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<150>;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly clientName: Varchar<150> | null;
      readonly status: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly TaskDependencies: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly dependentTaskId: CodecTypes['pg/uuid@1']['output'];
      readonly dependencyTaskId: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Tasks: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly projectId: CodecTypes['pg/uuid@1']['output'];
      readonly assignedToId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly title: Varchar<200>;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'TODO' | 'BLOCKED' | 'IN_PROGRESS' | 'DONE';
      readonly clientVisible: CodecTypes['pg/bool@1']['output'];
      readonly version: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<100>;
      readonly email: Varchar<255>;
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly role: 'PM' | 'INTERNAL' | 'CLIENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly AuditLogs: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly taskId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly changedColumn: CodecTypes['sql/varchar@1']['input'];
      readonly oldValue: CodecTypes['pg/text@1']['input'] | null;
      readonly newValue: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly ProjectMembers: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly projectId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Projects: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly clientName: CodecTypes['sql/varchar@1']['input'] | null;
      readonly status: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly TaskDependencies: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly dependentTaskId: CodecTypes['pg/uuid@1']['input'];
      readonly dependencyTaskId: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Tasks: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly projectId: CodecTypes['pg/uuid@1']['input'];
      readonly assignedToId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly title: CodecTypes['sql/varchar@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'TODO' | 'BLOCKED' | 'IN_PROGRESS' | 'DONE';
      readonly clientVisible: CodecTypes['pg/bool@1']['input'];
      readonly version: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly role: 'PM' | 'INTERNAL' | 'CLIENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly audit_logs: {
      readonly changed_column: Varchar<100>;
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly new_value: CodecTypes['pg/text@1']['output'] | null;
      readonly old_value: CodecTypes['pg/text@1']['output'] | null;
      readonly task_id: CodecTypes['pg/uuid@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly project_members: {
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly project_id: CodecTypes['pg/uuid@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly projects: {
      readonly client_name: Varchar<150> | null;
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deleted_at: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<150>;
      readonly status: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED';
      readonly updated_at: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly task_dependencies: {
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly dependency_task_id: CodecTypes['pg/uuid@1']['output'];
      readonly dependent_task_id: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly tasks: {
      readonly assigned_to_id: CodecTypes['pg/uuid@1']['output'] | null;
      readonly client_visible: CodecTypes['pg/bool@1']['output'];
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deleted_at: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly project_id: CodecTypes['pg/uuid@1']['output'];
      readonly status: 'TODO' | 'BLOCKED' | 'IN_PROGRESS' | 'DONE';
      readonly title: Varchar<200>;
      readonly updated_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly version: CodecTypes['pg/int4@1']['output'];
    };
    readonly users: {
      readonly created_at: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly email: Varchar<255>;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<100>;
      readonly password_hash: CodecTypes['pg/text@1']['output'];
      readonly role: 'PM' | 'INTERNAL' | 'CLIENT';
      readonly updated_at: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;